requirements.txt
anthropic
pdfminer.six
